Move createCard functions into characterCard and cardsList components
