+++ src/index.ts
@@ -1,3 +1,4 @@
+import { characters } from "./characters/data/characters.js";
 import createApp from "./components/app/createApp.js";
 
 const root = document.querySelector(".root");
@@ -6,5 +7,5 @@ if (!root) {
   throw new Error("Root does not exist");
 }
 
-const appContent = createApp();
+const appContent = createApp(characters);
 root.appendChild(appContent);
